backend/requirements.txt
anyio==3.7.1
api==0.0.7
APScheduler==3.10.4
autopep8==2.0.2
bcrypt==4.0.1
blinker==1.6.2
brotlipy==0.7.0
cached-property==1.5.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
deprecation==2.1.0
distlib==0.3.7
ecdsa==0.18.0
exceptiongroup==1.1.2
fastapi==0.99.1
fastapi-utils==0.2.1
filelock==3.12.2
Flask==2.3.3
fonttools==4.39.2
gotrue==1.3.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
mysql-connector-python==8.1.0
nodeenv==1.8.0
nose==1.3.7
numpy==1.24.2
packaging==23.0
pandas==1.5.3
passlib==1.7.4
pika==1.3.1
Pillow==9.4.0
platformdirs==3.10.0
postgrest==0.13.0
prisma==0.9.1
protobuf==4.21.12
pyaes==1.6.1
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.11
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-keycloak==3.0.0
python-multipart==0.0.6
python-telegram-bot==20.7
pytz==2022.7.1
PyYAML==5.4.1
realtime==1.0.2
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
storage3==0.6.1
StrEnum==0.4.15
supabase==2.0.3
supafunc==0.3.1
Telethon==1.33.1
tomlkit==0.12.1
typing_extensions==4.7.1
tzlocal==5.2
urllib3==2.0.2
uvicorn==0.22.0
virtualenv==20.24.3
websockets==11.0.3
Werkzeug==2.3.7
